proj-2-agent read data from two endpoints using FastAPI/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
anthropic>=0.28.0
python-dotenv>=1.0.0
